requirement.txt
flask
pillow
qrcode

